Uncertainty calculation of the probabilities and plots implemented, chebyshev implemented, triangular mesh implemented and it works.  EVERYTHING IS WORKING except the square mesh for the hex assemblies which I need to reproduce again

+++ Shynt/api_py/ResponseMatrix/matrix_utilities.py
@@ -85,7 +85,7 @@ def getBlockMatrix(matrixes):
       corner[1] + m_shape[1]
     ) # The final value of the corner gives the Big matrix shape
 
-  bigM = np.zeros(corner)
+  bigM = np.zeros(corner, dtype=np.float64)
   for m in range(numMatrixes):
     matrix = matrixes[m]
     corner = matrix_corners[m]
